- написан HomeFragment

+++ core/src/main/java/com/test/application/core/domain/Hotel.kt
@@ -4,7 +4,7 @@ data class Hotel (
     val id: Int,
     val name: String,
     val address: String,
-    val minimalPrice: Long?,
+    val minimalPrice: Int?,
     val priceForIt: String?,
     val rating: Double?,
     val ratingName: String?,
